documentation and clean up.
